Split strongs fn into new module, and improve error messages
